requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
ffmpeg-python==0.2.0
Flask==2.3.2
Flask-Cors==3.0.10
future==0.18.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
requests==2.30.0
six==1.16.0
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
